+++ spear/Acb.Spear.Contracts/Dtos/Account/AccountDto.cs
@@ -18,7 +18,7 @@ namespace Acb.Spear.Contracts.Dtos.Account
         public string Avatar { get; set; }
 
         ///<summary> 角色 </summary>
-        public short Role { get; set; }
+        public byte Role { get; set; }
 
         ///<summary> 创建时间 </summary>
         public DateTime CreateTime { get; set; }
